// src/app/core/models/document.ts
export class Document{
    id?: string;
    nomDocument?: string;
    objetDocument?: string;
    dateNotification?: string;
    typeDocument?: string;
    delaiReponse?: string;
    dateAlerte?: any;
    emails?: string;
    statut?: string;
    piece_jointe?: string;
    intitule: string;
    objetEmail: string;
    datePremierNotif: string;
    dateSecondNotif: string;
    etatNotif:string;
}